+++ CHANGELOG.md
@@ -3,9 +3,29 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [3.6.0] - 2021-07-19
+## [3.8.0] - 2021-07-23
 ### Changed
 - Hide Research Study step if server has no Research Study data.
+
+## [3.7.0] - 2021-07-23
+### Changed
+- Automatically fallback to single requests if batching is not supported.
+
+## [3.6.1] - 2021-07-22
+### Changed
+- Show Patient id (dbGaP)
+- Store visible table columns for each service base URL separately
+
+## [3.6.0] - 2021-07-22
+### Added
+- Limit list of resources to what has data (dbGaP).
+### Fixed
+- It was possible to switch to the View cohort step without
+  searching for Patients.
+
+## [3.5.1] - 2021-07-22
+### Fixed
+- README.md
 
 ## [3.5.0] - 2021-07-12
 ### Added